+++ NEWS.md
@@ -1,5 +1,6 @@
 # orgdata 1.2.2
 - Add unidentified municipalities as `xx99` to `geo_map()`. Unidentified municipalities do not exist from API. (#304)
+- Age group with `AgeCat` can use mix categories easily with `[x]`. See example in `find_age_category()`. (#305) 
 
 # orgdata 1.2.1
 - Fix bugs when writing and reading data from warehouse ie. marked the